feat(osparc): Implement batch run GUI and logic
Adds a new GUI for running and monitoring oSPARC batch simulations. Includes main logic for job submission, polling, and result download. Closes #11
