+++ .bashrc
@@ -1,2 +1,1 @@
-#export PATH="/C/Program Files/Sim4Life_8.2.2.18061/Python:$PATH"
 export PATH="/C/Program Files/Sim4Life_8.2.0.16876/Python:$PATH"
