+++ configs/far_field_config.json
@@ -8,7 +8,6 @@
         "do_extract": true,
         "batch_run": false
     },
-    "manual_isolve": true,
     "phantoms": [
         "duke",
         "ella",
