+++ configs/profiling_config.json
@@ -1,7 +1,7 @@
 {
     "near_field": {
         "avg_setup_time": 158.76500000001397,
-        "avg_setup_simulation": 105.50266666666721,
+        "avg_setup_simulation": 27.95299999997951,
         "avg_extract_time": 15.405999999959022,
         "avg_run_write_input_file": 21.25800000000163,
         "avg_run_isolve_execution": 6.648499999981141,
@@ -13,11 +13,14 @@
         "avg_extract_point_sensor_data": 0.06300000002374873
     },
     "far_field": {
-        "avg_setup_simulation": 77.79129411764563,
-        "avg_setup_time": 1143.3589999999967,
-        "avg_extract_input_power": 0.015666666667433066,
-        "avg_extract_sar_statistics": 0.189999999999903,
-        "avg_extract_point_sensor_data": 0.5780833333337796,
-        "avg_extract_time": 46.84399999999732
+        "avg_setup_simulation": 42.40600000001723,
+        "avg_setup_time": 52.875,
+        "avg_extract_input_power": 0.125,
+        "avg_extract_sar_statistics": 11.547000000020489,
+        "avg_extract_point_sensor_data": 0.1720000000204891,
+        "avg_extract_time": 35.14100000000326,
+        "avg_run_write_input_file": 12.14000000001397,
+        "avg_run_simulation_total": 139.625,
+        "avg_run_isolve_execution": 10.953000000037719
     }
 }
